+++ packages/faro-bundlers-shared/src/index.ts
@@ -182,28 +182,4 @@ export const exportBundleIdToEnv = (bundleId: string, appName: string, verbose?:
   }
 };
 
-/**
- * Recursively finds all .map files in a directory
- * @param dir The directory to search
- * @returns An array of paths to all .map files in the directory and its subdirectories
- */
-export const findMapFiles = (dir: string) => {
-  const sourcemapFiles: string[] = [];
-  const files = fs.readdirSync(dir);
-
-  for (const file of files) {
-    const filePath = path.join(dir, file);
-    const stat = fs.statSync(filePath);
-
-    if (stat.isDirectory()) {
-      // Recursively search subdirectories
-      findMapFiles(filePath);
-    } else if (stat.isFile() && file.endsWith('.map')) {
-      sourcemapFiles.push(filePath);
-    }
-  }
-
-  return sourcemapFiles;
-};
-
 crypto.randomUUID();
