+++ packages/faro-cli/package.json
@@ -22,7 +22,7 @@
   },
   "dependencies": {
     "@grafana/faro-bundlers-shared": "^0.1.2",
-    "commander": "^11.1.0",
+    "commander": "^13.1.0",
     "jest": "^29.7.0",
     "tar": "^7.1.0"
   },
